qcwcn: Address format compilation issues

+++ qcwcn/wifi_hal/wifi_hal.cpp
@@ -3478,7 +3478,7 @@ char *get_iface_mask_str(u32 mask, char *buf, size_t buflen) {
     return buf;
 
 error:
-    ALOGE("snprintf() error res=%d, write length=%d", res, end - pos);
+    ALOGE("snprintf() error res=%d, write length=%d", res, static_cast<int>(end - pos));
     return NULL;
 }
